Rename files and packages and DRY composables
