+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.Xently.NoActionBar">
             <intent-filter>
                 <action android:name="co.ke.xently.action.MAIN" />
