+++ feature/shoppinglist/src/main/res/values/strings.xml
@@ -21,4 +21,5 @@
     <string name="fsl_text_field_label_purchase_quantity">Purchase quantity</string>
     <string name="fsl_recommendations_toolbar_title">Recommendations</string>
     <string name="drawer_menu_signout">Signout</string>
+    <string name="fsl_group_menu_update">Update</string>
 </resources>
