+++ feature/shops/src/main/java/co/ke/xently/shops/ui/detail/ShopDetailScreen.kt
