+++ feature/src/main/res/values/strings.xml
@@ -3,6 +3,8 @@
     <string name="default_currency" translatable="false">KES</string>
     <string name="drawer_menu_shops">Shops</string>
     <string name="drawer_menu_products">Products</string>
+    <string name="generic_error_message">Sorry, an unexpected error was encountered.</string>
+    <string name="empty_list">No %1$s added, yet!</string>
     <string name="stop_location_tracking_button_text">Stop tracking</string>
     <string name="show_screen_button_text">Show screen</string>
     <string name="location_tracking_notification_text">Tracking your location for better shop recommendation…</string>
